# api/docs/swagger.yaml
basePath: /
definitions:
...
  handler.TaskCompletionResponse:
    properties:
      message:
        example: Task completed successfully
        type: string
      task_id:
        example: 123
        type: integer
      xp_earned:
        example: 100
        type: integer
    type: object
  handler.TaskCreateRequest:
    properties:
      description:
        example: Нужно завершить все задачи по проекту до конца недели
        type: string
      due_date:
        example: "2024-01-15T00:00:00Z"
        type: string
      tag_id:
        example: 1
        type: integer
      title:
        example: Завершить проект
        type: string
      xp_reward:
        example: 100
        type: integer
    type: object
  handler.TaskUpdateRequest:
    properties:
      description:
        example: Обновленное описание задачи
        type: string
      due_date:
        example: "2024-01-20T00:00:00Z"
        type: string
      tag_id:
        example: 2
        type: integer
      title:
        example: Завершить проект - обновлено
        type: string
      xp_reward:
        example: 150
        type: integer
    type: object
...
  model.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      difficulty:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_done:
        type: boolean
      repeat_interval:
        type: string
      tag_id:
        type: integer
      title:
        type: string
      type:
        type: string
      user_id:
        type: integer
      xp_reward:
        type: integer
    type: object
...
host: 10.155.36.40:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для FarmFocus - системы управления задачами и фермой
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FarmFocus API
  version: "1.0"
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: Возвращает список всех задач текущего пользователя
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить все задачи пользователя
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создает новую задачу для текущего пользователя
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Данные для создания задачи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.TaskCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новую задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет задачу по ID
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить задачу
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Возвращает задачу по указанному ID
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить задачу по ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Обновляет информацию о задаче
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные задачи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.TaskUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить задачу
      tags:
      - tasks
  /tasks/{id}/done:
    patch:
      consumes:
      - application/json
      description: Помечает задачу как выполненную, добавляет опыт пользователю и
        создает запись в логе прогресса
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TaskCompletionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Пометить задачу как выполненную
      tags:
      - tasks
  /tasks/{id}/undone:
    patch:
      consumes:
      - application/json
      description: Помечает задачу как невыполненную (сбрасывает статус)
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Пометить задачу как невыполненную
      tags:
      - tasks
  /tasks/overdue:
    get:
      consumes:
      - application/json
      description: Возвращает список просроченных задач (с истекшим сроком выполнения)
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить просроченные задачи
      tags:
      - tasks
  /tasks/status:
    get:
      consumes:
      - application/json
      description: Возвращает задачи по статусу выполнения (pending или completed)
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Статус задач (pending/completed)
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить задачи по статусу
      tags:
      - tasks
  /tasks/tag/{tagId}:
    get:
      consumes:
      - application/json
      description: Возвращает задачи, привязанные к определенному тегу
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Tag ID
        in: path
        name: tagId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить задачи по тегу
      tags:
      - tasks
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"